4. API Test로 변경
